Rename array.rs to arrays.rs (and same for array_const_generics.rs)
