+++ tests/Makefile
@@ -25,7 +25,7 @@ charon-tests: \
 	test-loops test-loops_cfg test-hashmap \
 	test-paper test-hashmap_main \
 	test-matches test-matches_duplicate test-external \
-	test-constants test-array test-array_const_generics test-traits \
+	test-constants test-arrays test-arrays_const_generics test-traits \
 	test-closures test-bitwise
 
 test-nested_borrows: OPTIONS += --no-code-duplication
@@ -40,8 +40,8 @@ test-constants: OPTIONS += --no-code-duplication
 test-matches:
 test-external: OPTIONS += --no-code-duplication
 test-matches_duplicate:
-test-array:
-test-array_const_generics:
+test-arrays:
+test-arrays_const_generics:
 test-traits:
 test-traits_special:
 test-closures:
