+++ tests/src/lib.rs
@@ -1,5 +1,5 @@
-pub mod array;
-pub mod array_const_generics;
+pub mod arrays;
+pub mod arrays_const_generics;
 pub mod bitwise;
 pub mod closures;
 pub mod constants;
